requirements-web.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pyyaml==6.0.1
pytest==8.2.0
httpx==0.27.0
